+++ include/runtime_api.h
@@ -4,7 +4,7 @@
 extern "C" {
 
 // Allocate a CUDA tensor
-Tensor* rt_alloc(int rank, const int* shape, size_t element_size);
+Tensor* rt_alloc(int rank, const int64_t* shape, size_t element_size);
 
 // Free tensor
 void rt_free(Tensor* t);
