+++ src/runtime_api.cpp
@@ -6,8 +6,8 @@ static Allocator g_allocator;
 
 extern "C" {
 
-Tensor* rt_alloc(int rank, const int* shape, size_t element_size) {
-    std::vector<int> dims(shape, shape + rank);
+Tensor* rt_alloc(int rank, const int64_t* shape, size_t element_size) {
+    std::vector<int64_t> dims(shape, shape + rank);
     Tensor* t = new Tensor(
         g_allocator.allocate(dims, element_size, DeviceType::CUDA)
     );
